add ability to use object properties as function or method arguments
The object itself is what will be used for dependency verification not
the individual properties, they will be assumed to be there. Additionally
callbacks will not be able to write to object properties to make the
dependencies clear.

So for arguments, you may use obj, obj.foo, or obj.foo.bar etc.