refactor to use inputParser. Add execOptions
Execution options can be passed in as objects on the input args,
they simply have to have a truthy property reactExecOptions and
they will be merged with the default exec options from left to
right. This allows them to be bound to a function, exposing a new
style function like promises, event emitter, etc.

The exec options will be pulled out of the argument list before
the arguments are assigned to their values and they can be
anywhere in the list.